refactor(storage): inline object metadata read entrypoints
